requirements.txt
numpy==2.1.2
PyAutoGUI==0.9.54
pylsl==1.16.2
PyYAML==6.0.2
